+++ src/me/dalton/capturethepoints/util/ArenaUtils.java
@@ -87,7 +87,7 @@ public class ArenaUtils {
      * @param player The player to heal.
      * @param amount The amount to heal the player.
      */
-    public void healPlayerAndCallEvent(Player player, double amount) {
+    public void healPlayerAndCallEvent(Player player, int amount) {
     	player.setHealth(amount);
     	EntityRegainHealthEvent regen = new EntityRegainHealthEvent(player, amount, RegainReason.CUSTOM);
     	ctp.getPluginManager().callEvent(regen);
@@ -100,7 +100,7 @@ public class ArenaUtils {
      * @param player The player to heal.
      */
     public void setFullHealthPlayerAndCallEvent(Arena arena, Player player) {
-    	double gained = player.getMaxHealth() - player.getHealth();
+    	int gained = player.getMaxHealth() - player.getHealth();
     	
     	player.setHealth(player.getMaxHealth());
